Mandelbrot set
The Mandelbrot set is the set of complex numbers c for which the function fc(z)=z^2+c does not diverge when iterated from z=0, i.e., for which the sequence fc(0), fc(fc(0)), etc., remains bounded in absolute value. Its definition and name are due to Adrien Douady, in tribute to the mathematician Benoit Mandelbrot.