Code for Julia set derived from the Mandelbrot set
In the context of complex dynamics, a topic of mathematics, the Julia set and the Fatou set are two complementary sets (Julia 'laces' and Fatou 'dusts') defined from a function. Informally, the Fatou set of the function consists of values with the property that all nearby values behave similarly under repeated iteration of the function, and the Julia set consists of values such that an arbitrarily small perturbation can cause drastic changes in the sequence of iterated function values. Thus the behavior of the function on the Fatou set is 'regular', while on the Julia set its behavior is 'chaotic'.

The Julia set of a function f is commonly denoted J(f), and the Fatou set is denoted F(f).[1] These sets are named after the French mathematicians Gaston Julia[2] and Pierre Fatou[3] whose work began the study of complex dynamics during the early 20th century.