Genetic algorithm - rockets flying towards target
Based on Jer Thorp's idea, this genetic algorithm is a simulation of rockets flying towards the target, learning along the way picking the best optimal route based on the distance between them and the target and whether they have crashed or not. The first population gets a random vector so a lot of them dies at the bottom of the screen, but they learn from there.