Calculating collisions with quadtree
A quadtree is a tree data structure in which each internal node has exactly four children. Quadtrees are the two-dimensional analog of octrees and are most often used to partition a two-dimensional space by recursively subdividing it into four quadrants or regions. The data associated with a leaf cell varies by application, but the leaf cell represents a "unit of interesting spatial information".

Quadtree can be used to optimise algorithms like calculating object collisions, by using a range to determine a handful of objects that might be colliding instead of checking object collision against every other object.